bug fixed

+++ workspace/turtlebot3/clean_build.bash
@@ -4,6 +4,12 @@ rm -rf build
 rm -rf devel
 
 DIR_PATH=$(cd ../../third-party/ROS-TCP-Endpoint && pwd)
-ln -s ${DIR_PATH} src/ROS-TCP-Endpoint
+
+if [ -f src/ROS-TCP-Endpoint ]
+then
+	:
+else
+	ln -s ${DIR_PATH} src/ROS-TCP-Endpoint
+fi
 
 catkin_make
